install: Make sure old symlinks are removed before creating a new one
Ignore .zsh_dotfiles/.zsh_secrets since a symlink is created here

+++ install_symlinks.sh
@@ -5,33 +5,48 @@ DOTFILES_PATH=${script_path%/*}
 
 symlink_to_home() {
   local_path="$1"
+  destination_path="$HOME/$1"
 
-  rm "$HOME/$1"
-  ln -s "$DOTFILES_PATH/$1" "$HOME/$1"
+  if [ -h "$destination_path" ]; then
+     rm "$HOME/$1"
+  fi
+
+  ln -s "$DOTFILES_PATH/$1" "$destination_path"
 }
 
 # nvim
+# ----
 symlink_to_home ".config/nvim"
 
 # General scripts on $PATH
+# ------------------------
 mkdir -p "$HOME/.local/bin"
 symlink_to_home ".local/bin/parallel-commands"
 
 # Tmux
+# ----
 symlink_to_home ".local/bin/tmux-sessionizer"
 symlink_to_home ".tmux.conf"
 
 # zsh
-if [ ! -d "$HOME/.z_dotfiles" ]; then
-  mkdir "$HOME/.z_dotfiles"
+# ---
+
+# z
+if [ -d "$HOME/.z_dotfiles" ]; then
+  rm -r "$HOME/.z_dotfiles"
 fi
-rm "$HOME/.z_dotfiles/.z_secrets"
+mkdir "$HOME/.z_dotfiles"
 ln -s "$DOTFILES_PATH/secrets/.z_secrets" "$HOME/.z_dotfiles/.z_secrets"
 
+# zsh
 symlink_to_home ".zsh_dotfiles"
+if [ -h "$HOME/.zsh_dotfiles/.zsh_secrets" ]; then
+  rm "$HOME/.zsh_dotfiles/.zsh_secrets"
+fi
 ln -s "$DOTFILES_PATH/secrets/.zsh_secrets" "$HOME/.zsh_dotfiles/.zsh_secrets"
 
 # Single files
+# ------------
 symlink_to_home ".gitignore_global"
 symlink_to_home ".ideavimrc"
 symlink_to_home ".gitconfig"
